Dashboard as the root page
To fix login page layout
