--- app/dashboard/index.ts
@@ -1,3 +1,0 @@
-import MainContent from "./MainContent";
-
-export { MainContent };
