+++ app/layout.tsx
@@ -1,27 +1,24 @@
-import { Navigation } from "./components";
-import "./globals.css";
+import { Navigation } from "./components"
+import "./globals.css"
 
 export const metadata = {
   title: "Smart Hard Hat System",
   description: "Web application for administrator usage",
-};
+}
 
 export default function RootLayout({
   children,
 }: {
-  children: React.ReactNode;
+  children: React.ReactNode
 }) {
+
   return (
     <html lang="en">
-      {/*
-        <head /> will contain the components returned by the nearest parent
-        head.tsx. Find out more at https://beta.nextjs.org/docs/api-reference/file-conventions/head
-      */}
       <head />
       <body className="flex">
         <Navigation />
         {children}
       </body>
     </html>
-  );
+  )
 }
