+++ app/login/index.ts
@@ -1,0 +1,3 @@
+import LoginComponent from "./LoginComponent"
+
+export { LoginComponent }
