+++ app/login/layout.tsx
@@ -1,0 +1,11 @@
+export default function LoginLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return (
+      <div>
+          {children}
+      </div>
+  )
+}
