+++ app/login/loading.tsx
@@ -1,0 +1,7 @@
+export default function Loading() {
+  return (
+    <>
+      <h1>Loading...</h1>
+    </>
+  )
+}
