+++ app/login/page.tsx
@@ -1,0 +1,9 @@
+import { LoginComponent } from "."
+
+export default function Home() {
+  return (
+    <main className="min-h-screen w-screen grid place-content-center">
+      <LoginComponent />
+    </main>
+  )
+}
